# optimized brigand::partition