```cpp
template <typename N>
using is_odd = std::integral_constant<bool, (N::value % 2)>;

using list = brigand::partition<brigand::range<int, 0, SIZE>, brigand::quote<is_odd>>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.09s - 38M | 00.12s - 41M | 00.20s - 48M | 00.41s - 72M
gcc | 00.05s - 21M | 00.08s - 26M | 00.13s - 33M | 00.23s - 51M | 00.49s - 98M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.09s - 38M | 00.12s - 40M | 00.19s - 45M | 00.34s - 57M
gcc | 00.05s - 21M | 00.08s - 26M | 00.12s - 32M | 00.20s - 45M | 00.38s - 71M